Move BagId out of BagCourier
