+++ lib/bag_id.rb
@@ -1,0 +1,18 @@
+
+module BagId
+  class BagId
+    attr_reader :repository, :object_id, :context, :part_id
+
+    def initialize(repository:, object_id:, context: nil, part_id: nil)
+      @repository = repository
+      @context = context
+      @object_id = object_id
+      @part_id = part_id
+    end
+
+    def to_s
+      segments = [@context, @object_id, @part_id].select { |segment| !segment.nil? }
+      "#{@repository}.#{segments.join "-"}"
+    end
+  end
+end
